fucked up. Should be floats not doubles

+++ c/test/main.c
@@ -3,14 +3,14 @@
 
 int main (int argc, const char* argv[])
 {
-	double* in = (double*) malloc(sizeof(double) * 1024*1024*1024);
-	double* out = (double*) malloc(sizeof(double) * 1024 * 1024 * 1024);
-	double t = omp_get_wtime();
+	float* in = (float*) malloc(sizeof(float) * 1024*1024*1024);
+	float* out = (float*) malloc(sizeof(float) * 1024 * 1024 * 1024);
+	float t = omp_get_wtime();
 	for (int i = 0; i < 5; i++)
 	{
 		apply(in, out);
 	}
-	double t2 = omp_get_wtime();
+	float t2 = omp_get_wtime();
 	free(in);
 	free(out);
 	printf("Time: %f\n", t2-t);
